requirements.txt
finnhub_python==2.4.25
pandas==2.3.3
plotly==6.3.1
psycopg2_binary==2.9.11
python-dotenv==1.1.1
SQLAlchemy==2.0.44
streamlit==1.50.0
streamlit_autorefresh==1.0.1
websocket_client==1.9.0
